requirements-api.txt
# RapidDoc API Server 依赖包
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Logging and utilities
loguru==0.7.2
python-json-logger==2.0.7

# File processing
pypdfium2==0.73.0
Pillow==10.1.0
opencv-python==4.8.1.78

# OCR and AI models
rapidocr-openvino==1.3.0
onnxruntime==1.16.3
torch>=2.0.0
torchvision>=0.15.0

# Math and formula processing
latex2mathml==3.76.0
rapidlatex==0.2.1

# Data processing
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# HTTP and async
aiofiles==23.2.1
httpx==0.25.2

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0